requirements.txt
docker==7.1.0
fastapi-slim==0.111.0
fastapi-cli==0.0.4
uvicorn==0.29.0
prometheus-client==0.20.0
